Refactor the entire code practice two
