+++ practice-one/src/components/common/Card/index.tsx
@@ -29,7 +29,6 @@ type CustomCardProps = {
   onSubmit?: () => void;
 };
 
-// Component Card
 const Card: React.FC<CustomCardProps> = memo(
   ({
     className,
