+++ practice-one/src/components/common/Error/index.tsx
@@ -6,7 +6,6 @@ type TError = {
   content: string;
 };
 
-//Component Error
 const Error: React.FC<TError> = memo(
   ({ className, content }): React.ReactElement => {
     return <p className={className}>{content}</p>;
