+++ practice-one/src/components/common/Loading/index.tsx
@@ -1,11 +1,12 @@
+// React
+import React from "react";
+
 // Image
 import { Spinner } from "../../../assets/image";
 
 // Component
 import { Image } from "..";
-import React from "react";
 
-// Component Loading
 const Loading: React.FC = (): React.ReactElement => {
   return (
     <div className="spinner">
