+++ practice-one/src/helpers/CalculatorToltalPrice.ts
@@ -4,7 +4,6 @@ import { CustomProductProps } from "../types/Product";
 // Calculate total value
 export const calculatorTotalPrice = (carts: CustomProductProps[]): number => {
   if (!carts) return 0;
-
   const totalPrice = carts.reduce(
     (accumulator: number, item: CustomProductProps) => {
       const itemTotal = (item.quantity || 0) * (item.price || 0);
