+++ practice-one/src/helpers/FetchData.ts
@@ -1,10 +1,9 @@
+import { ERROR_MESSAGES } from "../constants/Error";
+
 /**
  * FetchData function to fetch data from URL and process results
  * @param url
  */
-
-import { ERROR_MESSAGES } from "../constants/Error";
-
 export const fetchData = async (url: string) => {
   const response = await fetch(url);
   if (response.ok) {
