+++ practice-one/src/pages/Checkout/index.tsx
@@ -4,7 +4,6 @@ import React from "react";
 // Component
 import { FormCheckout } from "../../components";
 
-// Component Checkout
 const Checkout: React.FC = (): React.ReactElement => {
   return <FormCheckout />;
 };
