+++ practice-one/src/pages/Home/index.tsx
@@ -1,14 +1,13 @@
 // CSS
 import "./Home.css";
 
+// React
+import React from "react";
+import { Link } from "react-router-dom";
+
 // Component
 import { Heading } from "../../components";
 
-// React router
-import { Link } from "react-router-dom";
-import React from "react";
-
-// Component HomePage
 const Home: React.FC = (): React.ReactElement => {
   return (
     <div className="home-layout font-family">
