+++ practice-one/src/pages/Menu/index.tsx
@@ -7,7 +7,6 @@ import ErrorBoundary from "../../components/ErrorBoundary";
 
 const ListProduct = lazy(() => import("../../components/ListProduct/index"));
 
-// Component Menu
 const Menu: React.FC = (): React.ReactElement => {
   return (
     <Suspense fallback={<Loading />}>
